# dsa-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the DSA Backend application.",
        "title": "DSA Backend API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/user/login": {
            "post": {
                "description": "User login with user ID and password. Returns a JWT token if successful.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful. Returns a JWT token.",
                        "schema": {
                            "$ref": "#/definitions/user.userLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request. This error occurs if the user ID or password is missing or incorrect.",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error. This error occurs if there is an issue with the database or password hashing.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": [
                            "me"
                        ]
                    }
                ],
                "description": "Logout user and invalidate JWT token",
                "tags": [
                    "User"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/me": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "me"
                        ]
                    }
                ],
                "description": "Get current user information from JWT token",
                "tags": [
                    "User"
                ],
                "summary": "Get current user information",
                "responses": {
                    "200": {
                        "description": "Current user information",
                        "schema": {
                            "$ref": "#/definitions/user.userResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.Error": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "properties": {
                        "body": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "response.Success": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "user.userLoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "DO NOT modify json name, 'access_token' is required in Swagger UI",
                    "type": "string"
                },
                "exp": {
                    "type": "integer"
                },
                "token_type": {
                    "description": "DO NOT modify json name, 'token_type' is required in Swagger UI",
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/user.userResponse"
                }
            }
        },
        "user.userResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/user/login",
            "scopes": {
                "admin": "Grants any rights related to admin user",
                "manager": "Grants any rights related to manager user",
                "me": "Grants any rights related to the current user"
            }
        }
    }
}